switch to httpx code cleanup
