+++ notifications_android_tv/exceptions.py
@@ -1,0 +1,7 @@
+"""Exceptions raised by the library."""
+
+class ConnectError(Exception):
+    """Exception raised for connection error."""
+
+class InvalidResponse(Exception):
+    """Exception raised for invalid response."""
